Write software block styles
I didn't use the <progress> element because I think it's more suited for
cases where there's actual progress happening (e.g., file uploads). And
it looks like styling them is a pain, with lots of browser-specific
stuff.

On the other hand, this is confusing for screen reader users. Though I'm
not sure if actual <progress> bars are read by screen readers.

For some reason, when the height of the progress bars is .4rem, they
don't look like the same height. Probably because of rounding off the
pixels. Meanwhile .5rem looks fine (because it gives a nice whole number
when converted to pixels). Maybe I should just give it plain pixel
values.

I need the set the width for the table headers to something like 35%,
because the progress bars (being empty divs themselves) would collapse
to zero-width. Might as well do the same with the earlier table, so they
line up nicely.