Make flex-list block
This is for the nav bar list and the social media icons list.

To make the block applicable for both lists, I had to wrap the social
list in a div, then make that the social and row mix, so it's the same as the
nav bar (where the list is nested in a row).